Refactor and add new entities and properties
Minor cleanup in ChangeStatusApplicationForm.cs by removing an unnecessary character from the using statement.

In ImportExcel.cs, changed the ImportExcel command to return a string instead of a List<string>. Updated ImportExcelHandler to include a new dependency IDepartmentRepository, fetch manager information, and set ManagerId for Employee. Changed the return type of the Handle method to Result<string>.

Added new properties HeadOfDepartmentId and HeadOfDepartment to the Department class in Department.cs to link it with the AppUser entity.

Added a new property NumberOfLeave to the AppUser class in AppUser.cs.

Added a new DbSet<AbsentForm> to the ApplicationDbContext class in ApplicationDbContext.cs.

Introduced a new entity class AbsentForm in AbsentForm.cs, inheriting from BaseForm and including From and To date properties.

Added a new repository interface IAbsentFormRepository in IAbsentFormRepository.cs.

Implemented the IAbsentFormRepository interface in the new repository class AbsentFormRepository in AbsentFormRepository.cs.